+++ contracts/V1/StateChannelDiamondProxy/StateChannelManagerStorage.sol
@@ -6,12 +6,11 @@ import "../AStateMachine.sol";
 
 contract StateChannelManagerStorage {
     //Config shared across all instances of the state machine
-    uint public p2pTime;
-    uint public agreementTime;
-    uint public chainFallbackTime;
-    uint public challengeTime;
+    uint256 public p2pTime;
+    uint256 public agreementTime;
+    uint256 public chainFallbackTime;
+    uint256 public challengeTime;
     uint256 public gasLimit;
-
 
     AStateMachine stateMachineImplementation;
 
@@ -23,27 +22,28 @@ contract StateChannelManagerStorage {
 
     /// @dev stateSnapshot Data
     mapping(bytes32 channelId => StateSnapshot) stateSnapshots;
-   
+
     // =================== Block on chain storage ==================
 
     /// @notice BlockCalldata Commitment
     /// @dev blockCalldataCommitments[channelId][signerAddress][forkCnt][blockHeight] => hash(off-chain block, on-chain block.timestamp)
-    mapping(bytes32 channelId => mapping(address signerAddress => mapping(uint forkCnt => mapping(uint blockHeight => bytes32 blockCallDataCommitment)))) blockCalldataCommitments;
-
+    mapping(
+        bytes32 channelId
+            => mapping(
+                address signerAddress
+                    => mapping(uint256 forkCnt => mapping(uint256 blockHeight => bytes32 blockCallDataCommitment))
+            )
+    ) blockCalldataCommitments;
 
     // ================== Dispute on chain storage ==================
 
     /// @dev disputeData[channelId] => DisputeData
     mapping(bytes32 channelId => DisputeData) disputeData;
 
-
     // ================== Modifiers ==================
 
     modifier onlySelf() {
-        require(
-            address(this) == msg.sender,
-            "Only self (facet) can call this (diamond) function"
-        );
+        require(address(this) == msg.sender, "Only self (facet) can call this (diamond) function");
         _;
     }
 }
