+++ contracts/V1/StateChannelManagerInterface.sol
@@ -4,89 +4,67 @@ import "./DataTypes.sol";
 import "./DisputeTypes.sol";
 
 abstract contract StateChannelManagerInterface {
-    function openChannel(
-        bytes32 channelId,
-        bytes[] calldata openChannelData,
-        bytes[] calldata signatures
-    ) public virtual;
+    function openChannel(bytes32 channelId, bytes[] calldata openChannelData, bytes[] calldata signatures)
+        public
+        virtual;
 
-    function closeChannel(
-        bytes32 channelId,
-        bytes[] calldata closeChannelData,
-        bytes[] calldata signatures
-    ) public virtual;
+    function closeChannel(bytes32 channelId, bytes[] calldata closeChannelData, bytes[] calldata signatures)
+        public
+        virtual;
 
-    function removeParticipant(
-        bytes32 channelId,
-        bytes[] calldata removeParticipantData,
-        bytes[] calldata signatures
-    ) public virtual;
+    function removeParticipant(bytes32 channelId, bytes[] calldata removeParticipantData, bytes[] calldata signatures)
+        public
+        virtual;
 
-    function processExitChannel(
-        bytes32 channelId,
-        ExitChannel calldata exitChannel
-    ) public virtual;
+    function processExitChannel(bytes32 channelId, ExitChannel calldata exitChannel) public virtual;
 
-    function addParticipant(
-        bytes32 channelId,
-        bytes[] calldata removeParticipantData,
-        bytes[] calldata signatures
-    ) public virtual;
+    function addParticipant(bytes32 channelId, bytes[] calldata removeParticipantData, bytes[] calldata signatures)
+        public
+        virtual;
 
-    function isChannelOpen(
-        bytes32 channelId
-    ) public view virtual returns (bool);
+    function isChannelOpen(bytes32 channelId) public view virtual returns (bool);
 
-    function getForkCnt(bytes32 channelId) public view virtual returns (uint);
+    function getForkCnt(bytes32 channelId) public view virtual returns (uint256);
 
-    function getParticipants(
-        bytes32 channelId
-    ) public virtual returns (address[] memory);
+    function getParticipants(bytes32 channelId) public virtual returns (address[] memory);
 
-    function getNextToWrite(
-        bytes32 channelId,
-        bytes memory encodedState
-    ) public virtual returns (address);
+    function getNextToWrite(bytes32 channelId, bytes memory encodedState) public virtual returns (address);
 
-    function getP2pTime() public view virtual returns (uint);
+    function getP2pTime() public view virtual returns (uint256);
 
-    function getAgreementTime() public view virtual returns (uint);
+    function getAgreementTime() public view virtual returns (uint256);
 
-    function getChainFallbackTime() public view virtual returns (uint);
+    function getChainFallbackTime() public view virtual returns (uint256);
 
-    function getChallengeTime() public view virtual returns (uint);
+    function getChallengeTime() public view virtual returns (uint256);
 
-    function getAllTimes() public view virtual returns (uint, uint, uint, uint);
+    function getAllTimes() public view virtual returns (uint256, uint256, uint256, uint256);
 
-    function getChainLatestBlockTimestamp(
-        bytes32 channelId,
-        uint forkCnt,
-        uint maxTransactionCnt
-    ) public view virtual returns (uint);
+    function getChainLatestBlockTimestamp(bytes32 channelId, uint256 forkCnt, uint256 maxTransactionCnt)
+        public
+        view
+        virtual
+        returns (uint256);
 
-    function executeStateTransitionOnState(
-        bytes32 channelId,
-        bytes memory encodedState,
-        Transaction memory _tx
-    ) public virtual returns (bool, bytes memory);
+    function executeStateTransitionOnState(bytes32 channelId, bytes memory encodedState, Transaction memory _tx)
+        public
+        virtual
+        returns (bool, bytes memory);
 
-    function postBlockCalldata(SignedBlock memory signedBlock, uint maxTimestamp) public virtual;
+    function postBlockCalldata(SignedBlock memory signedBlock, uint256 maxTimestamp) public virtual;
 
-    function getBlockCallDataCommitment(
-        bytes32 channelId,
-        uint forkCnt,
-        uint blockHeight,
-        address participant
-    ) public view virtual returns (bool found, bytes32 blockCalldataCommitment);
+    function getBlockCallDataCommitment(bytes32 channelId, uint256 forkCnt, uint256 blockHeight, address participant)
+        public
+        view
+        virtual
+        returns (bool found, bytes32 blockCalldataCommitment);
 
-    function createDispute(
-        Dispute memory dispute
-    ) public virtual;
+    function createDispute(Dispute memory dispute) public virtual;
 
-    function auditDispute(
-        Dispute memory dispute,
-        DisputeAuditingData memory disputeAuditingData 
-    ) public virtual returns (address[] memory slashParticipants);
+    function auditDispute(Dispute memory dispute, DisputeAuditingData memory disputeAuditingData)
+        public
+        virtual
+        returns (address[] memory slashParticipants);
 
     function challengeDispute(
         Dispute memory dispute,
